% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_hw.R
\name{check_hw}
\alias{check_hw}
\title{Check homework structure}
\usage{
check_hw(hw_id, netid, directory = ".")
}
\arguments{
\item{hw_id}{The name of the homework assignment, e.g. hw00}

\item{netid}{An alpha numeric character string not exceeding 9 letters.}
}
\value{
A `check_hw` object that contains a list of `TRUE` and `FALSE`
for each condition being observed.
}
\description{
Tests to see if the homework is structured properly.
}
\examples{
# Good
check_hw("hw00", "toad1")

# Bad
check_hw("hw00", "toad!!1")
}
