% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat385-package.R
\docType{package}
\name{stat385-package}
\alias{stat385}
\alias{stat385-package}
\title{stat385: What the Package Does (one line, title case)}
\description{
What the package does (one paragraph).
}
\details{
A companion package for STAT 385 @ UIUC
}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu}

}
