% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_hw_folder.R
\name{valid_hw_folder}
\alias{valid_hw_folder}
\title{Valid folder contents}
\usage{
valid_hw_folder(name)
}
\description{
Valid folder contents
}
\examples{
# Valid
valid_hw_folder("hw00")

# Not valid
valid_hw_folder("toad")
}
