% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_hw_files.R
\name{valid_hw_rmd_file}
\alias{valid_hw_rmd_file}
\title{Valid folder contents}
\usage{
valid_hw_rmd_file(name, folder)
}
\description{
Valid folder contents
}
\examples{
# Valid
valid_hw_rmd_file("hw00-netid.Rmd")

# Not valid
valid_hw_rmd_file("hw00-netid.Rmd")
}
